+++ src/ksp/src/DijkstraShortestPathAlg.h
@@ -34,7 +34,7 @@ class DijkstraShortestPathAlg: public Graph {
        ///@{
 
        //! Dijkstra using the original ids
-       std::deque<BaseEdge> Dijkstra(int source, int sink);
+       std::deque<BaseEdge> Dijkstra(long source, long sink);
        //! Dijkstra using the graph's IDs
        BasePath Dijkstra(UINT source, UINT sink);
        ///@}
